+++ packages/components/src/components.d.ts
@@ -6,7 +6,7 @@
  */
 import { HTMLStencilElement, JSXBase } from "@stencil/core/internal";
 import { Events, Props } from "./types";
-import { BalMode } from "./utils/config";
+import { BalConfigState, BalMode } from "./utils/config";
 import { BalCarouselItemData } from "./components/bal-carousel/bal-carousel.type";
 import { BalCheckboxOption } from "./components/form/bal-checkbox/bal-checkbox.type";
 import { Frameworks } from "./components/docs/bal-doc-stackblitz/stackblitz.util";
@@ -31,6 +31,7 @@ export namespace Components {
           * Label of the close trigger button
          */
         "closeLabel": string;
+        "configChanged": (state: BalConfigState) => Promise<void>;
         /**
           * Set the amount of time, in milliseconds, to wait to trigger the `balChange` event after each keystroke. This also impacts form bindings such as `ngModel` or `v-model`.
          */
@@ -560,6 +561,7 @@ export namespace Components {
           * Closes the datepicker popover after selection
          */
         "closeOnSelect": boolean;
+        "configChanged": (state: BalConfigState) => Promise<void>;
         /**
           * Set the amount of time, in milliseconds, to wait to trigger the `ionChange` event after each keystroke. This also impacts form bindings such as `ngModel` or `v-model`.
          */
@@ -802,6 +804,7 @@ export namespace Components {
         "subject"?: string;
     }
     interface BalFieldLabel {
+        "configChanged": (state: BalConfigState) => Promise<void>;
         /**
           * If `true`, the element is not mutable, focusable, or even submitted with the form. The user can neither edit nor focus on the control, nor its form control descendants.
          */
@@ -928,6 +931,7 @@ export namespace Components {
         "value": File[];
     }
     interface BalFooter {
+        "configChanged": (state: BalConfigState) => Promise<void>;
         /**
           * If `true` the language selection will be hidden.
          */
@@ -995,6 +999,7 @@ export namespace Components {
           * Text for the close button.
          */
         "closeLabel"?: string;
+        "configChanged": (state: BalConfigState) => Promise<void>;
         /**
           * Closes the hint box.
          */
@@ -1021,6 +1026,7 @@ export namespace Components {
           * The theme type of the button.
          */
         "color": Props.BalIconColor;
+        "configChanged": (state: BalConfigState) => Promise<void>;
         /**
           * If `true` the icon has display inline style
          */
@@ -1246,6 +1252,7 @@ export namespace Components {
         "value"?: string | number;
     }
     interface BalInputStepper {
+        "configChanged": (state: BalConfigState) => Promise<void>;
         /**
           * Set the amount of time, in milliseconds, to wait to trigger the `balChange` event after each keystroke. This also impacts form bindings such as `ngModel` or `v-model`.
          */
@@ -1322,6 +1329,7 @@ export namespace Components {
           * If `true` the list item shows that it is clickable
          */
         "clickable": boolean;
+        "configChanged": (state: BalConfigState) => Promise<void>;
         /**
           * If `true` the list item can be hovered
          */
@@ -1723,6 +1731,7 @@ export namespace Components {
         "color": Props.BalNotificationColor;
     }
     interface BalNumberInput {
+        "configChanged": (state: BalConfigState) => Promise<void>;
         /**
           * Set the amount of time, in milliseconds, to wait to trigger the `balChange` event after each keystroke. This also impacts form bindings such as `ngModel` or `v-model`.
          */