+++ packages/components/src/index.ts
@@ -27,6 +27,7 @@ export { initializeBaloiseDesignSystem } from './initialize'
 export * from './components/form/bal-select/utils/bal-option.util'
 export {
   waitForComponent,
+  waitForDesignSystem,
   shallowReady,
   deepReady,
   wait,